AHHHH

+++ Projekt/Square.cs
@@ -8,26 +8,26 @@ public class EnemySQ
   public int Waypoint = 0;
   public int target;
   public (int x, int y) targetPos;
-  public (int x, int y) Position;
+  public (float x, float y) Position;
   public (int x, int y) Directions;
   public Rectangle rect;
 }
 
 public class TOWER
 {
-    public int towerCounter = 0;
-    public int target;
-    public double distance;
-    public int attackduration = 0;
-    public int posX;
-    public int posY;
-    public int radius;
+  public int towerCounter = 0;
+  public int target;
+  public double distance;
+  public int attackduration = 0;
+  public int posX;
+  public int posY;
+  public int radius;
 
-    // Constructor som initierar 
-    public TOWER(int x, int y, int r)
-    {
-        posX = x;  
-        posY = y;  
-        radius = r;  
-    }
+  // Constructor som initierar 
+  public TOWER(int x, int y, int r)
+  {
+    posX = x;
+    posY = y;
+    radius = r;
+  }
 }
